requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
boto3==1.40.21
botocore==1.40.21
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-easymde==1.0.4
django-ratelimit==4.1.0
django-storages==1.14.6
env==0.1.0
Faker==37.6.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
pretend==1.0.9
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
ratelimit==2.2.1
requests==2.32.5
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
